+++ matching/src/main/java/com/ssafy/matching/repository/GatheringMemberRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface GatheringMemberRepository extends JpaRepository<GatheringMember, Integer> {
     GatheringMember save(GatheringMember memberGathering);
-    void deleteByGatheringIdAndMemberId(int gatheringId, int memberId);
+    void deleteByGatheringIdAndMemberId(int gatheringId, long memberId);
 }
