+++ matching/src/main/java/com/ssafy/matching/service/LiveServiceImpl.java
@@ -50,7 +50,7 @@ public class LiveServiceImpl implements LiveService {
     }
 
     @Override
-    public void leaveLive(int liveId, int memberId) {
+    public void leaveLive(int liveId, long memberId) {
         liveMemberRepository.deleteByLiveIdAndMemberId(liveId, memberId);
     }
 }
